chnagement import remote
