+++ Play.py
@@ -5,6 +5,7 @@ from random import *
 import socket
 import time
 import sys
+import remote_play
 
 
 def play (map_title, team_1, team_1_type, team_2, team_2_type):
@@ -1328,194 +1329,5 @@ def create_order(long, larg,  team, ships, units_stats,peaks) :
         instruction_str += element +' '
     
     return instruction_str
-
-def get_IP():
-    """Returns the IP of the computer where get_IP is called.
-    
-    Returns
-    -------
-    computer_IP: IP of the computer where get_IP is called (str)
-    
-    Notes
-    -----
-    If you have no internet connection, your IP will be 127.0.0.1.
-    This IP address refers to the local host, i.e. your computer.
-    
-    """   
-    
-    return socket.gethostbyname(socket.gethostname())
-
-def connect_to_player(player_id, remote_IP='127.0.0.1', verbose=False):
-    """Initialise communication with remote player.
-    
-    Parameters
-    ----------
-    player_id: player id of the remote player, 1 or 2 (int)
-    remote_IP: IP of the computer where remote player is (str, optional)
-    verbose: True only if connection progress must be displayed (bool, optional)
-    
-    Returns
-    -------
-    connection: sockets to receive/send orders (tuple)
-    
-    Notes
-    -----
-    Initialisation can take several seconds.  The function only
-    returns after connection has been initialised by both players.
-    
-    Use the default value of remote_IP if the remote player is running on
-    the same machine.  Otherwise, indicate the IP where the other player
-    is running with remote_IP.  On most systems, the IP of a computer
-    can be obtained by calling the get_IP function on that computer.
-        
-    """ 
-    
-    # init verbose display
-    if verbose:
-        print('\n-------------------------------------------------------------')
-        
-    # open socket (as server) to receive orders
-    socket_in = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-    socket_in.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1) # deal with a socket in TIME_WAIT state
-
-    if remote_IP == '127.0.0.1':
-        local_IP = '127.0.0.1'
-    else:
-        local_IP = get_IP()
-    local_port = 42000 + (3-player_id)
-    
-    try:
-        if verbose:
-            print('binding on %s:%d to receive orders from player %d...' % (local_IP, local_port, player_id))
-        socket_in.bind((local_IP, local_port))
-    except:
-        local_port = 42000 + 100+ (3-player_id)
-        if verbose:
-            print('   referee detected, binding instead on %s:%d...' % (local_IP, local_port))
-        socket_in.bind((local_IP, local_port))
-
-    socket_in.listen(1)
-    if verbose:
-        print('   done -> now waiting for a connection on %s:%d\n' % (local_IP, local_port))
-
-    # open client socket used to send orders
-    socket_out = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-    socket_out.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1) # deal with a socket in TIME_WAIT state
-    
-    remote_port = 42000 + player_id
-    
-    connected = False
-    msg_shown = False
-    while not connected:
-        try:
-            if verbose and not msg_shown:
-                print('connecting on %s:%d to send orders to player %d...' % (remote_IP, remote_port, player_id))
-            socket_out.connect((remote_IP, remote_port))
-            connected = True
-            if verbose:
-                print('   done -> now sending orders to player %d on %s:%d' % (player_id, remote_IP, remote_port))
-        except:
-            if verbose and not msg_shown:
-                print('   connection failed -> will try again every 100 msec...')
-            time.sleep(.1)
-
-            msg_shown = True
-            
-    if verbose:
-        print()
-
-    # accept connection to the server socket to receive orders from remote player
-    socket_in, remote_address = socket_in.accept()
-    if verbose:
-        print('now listening to orders from player %d' % (player_id))
-            
-    # end verbose display
-    if verbose:
-        print('\nconnection to remote player %d successful\n-------------------------------------------------------------\n' % player_id)
-
-    # return sockets for further use     
-    return (socket_in, socket_out)
-
-def disconnect_from_player(connection):
-    """End communication with remote player.
-    
-    Parameters
-    ----------
-    connection: sockets to receive/send orders (tuple)
-    
-    """
-    
-    # get sockets
-    socket_in = connection[0]
-    socket_out = connection[1]
-    
-    # shutdown sockets
-    socket_in.shutdown(socket.SHUT_RDWR)    
-    socket_out.shutdown(socket.SHUT_RDWR)
-    
-    # close sockets
-    socket_in.close()
-    socket_out.close()
-    
-def notify_remote_orders(connection, orders):
-    """Notifies orders of the local player to a remote player.
-    
-    Parameters
-    ----------
-    connection: sockets to receive/send orders (tuple)
-    orders: orders of the local player (str)
-        
-    Raises
-    ------
-    IOError: if remote player cannot be reached
-    
-    """
-     
-    # get sockets
-    socket_in = connection[0]
-    socket_out = connection[1]
-
-    # deal with null orders (empty string)
-    if orders == '':
-        orders = 'null'
-    
-    # send orders
-    try:
-        socket_out.sendall(orders.encode())
-    except:
-        raise IOError('remote player cannot be reached')
-
-def get_remote_orders(connection):
-    """Returns orders from a remote player.
-
-    Parameters
-    ----------
-    connection: sockets to receive/send orders (tuple)
-        
-    Returns
-    ----------
-    player_orders: orders given by remote player (str)
-
-    Raises
-    ------
-    IOError: if remote player cannot be reached
-            
-    """
-   
-    # get sockets
-    socket_in = connection[0]
-    socket_out = connection[1]
-
-    # receive orders    
-    try:
-        orders = socket_in.recv(65536).decode()
-    except:
-        raise IOError('remote player cannot be reached')
-        
-    # deal with null orders
-    if orders == 'null':
-        orders = ''
-        
-    return orders
 
 play('fichier', sys.argv[1],sys.argv[2],sys.argv[3],sys.argv[4])
